Update 04-afsc.R
This should be about right, but needs a little more testing after (hopefully) Monday when the 2024 data will be added to the FOSS API pull just to triple-check that everything is the same between the data pulls. I will push a commit one  last time before the pull request is merged. Note that the FOSS pull is within a if(FALSE) statement so the script will work as it did before if you don't like this new way of pulling data. The benefit to pulling data through FOSS is that the package builder does not need to have access inside the AFSC firewall to access data. Tagging @Lewis-Barnett-NOAA for awareness.